+++ src/Storage/Storage.Management/File/GetAzureStorageShare.cs
@@ -16,6 +16,7 @@ using Microsoft.Azure.Commands.Management.Storage.Models;
 using Microsoft.Azure.Management.Internal.Resources.Utilities.Models;
 using Microsoft.Azure.Management.Storage;
 using Microsoft.Azure.Management.Storage.Models;
+using Microsoft.Rest.Azure;
 using System;
 using System.Collections.Generic;
 using System.Management.Automation;
@@ -26,20 +27,34 @@ namespace Microsoft.Azure.Commands.Management.Storage
     public class GetAzureStorageShareCommand : StorageFileBaseCmdlet
     {
         /// <summary>
-        /// AccountName Parameter Set
+        /// AccountName list Parameter Set
         /// </summary>
         private const string AccountNameParameterSet = "AccountName";
 
         /// <summary>
-        /// Account object parameter set 
+        /// Account object list parameter set 
         /// </summary>
         private const string AccountObjectParameterSet = "AccountObject";
+        /// <summary>
+        /// AccountName Parameter Set
+        /// </summary>
+        private const string AccountNameSingleParameterSet = "AccountNameSingle";
+
+        /// <summary>
+        /// Account object parameter set 
+        /// </summary>
+        private const string AccountObjectSingleParameterSet = "AccountObjectSingle";
 
         /// <summary>
         /// Share ResourceId  parameter set 
         /// </summary>
         private const string ShareResourceIdParameterSet = "ShareResourceId";
 
+        [Parameter(
+            Position = 0,
+            Mandatory = true,
+            HelpMessage = "Resource Group Name.",
+            ParameterSetName = AccountNameSingleParameterSet)]
         [Parameter(
             Position = 0,
             Mandatory = true,
@@ -52,11 +67,20 @@ namespace Microsoft.Azure.Commands.Management.Storage
             Position = 1,
             Mandatory = true,
             HelpMessage = "Storage Account Name.",
+            ParameterSetName = AccountNameSingleParameterSet)]
+        [Parameter(
+            Position = 1,
+            Mandatory = true,
+            HelpMessage = "Storage Account Name.",
             ParameterSetName = AccountNameParameterSet)]
         [Alias(AccountNameAlias)]
         [ValidateNotNullOrEmpty]
         public string StorageAccountName { get; set; }
 
+        [Parameter(Mandatory = true,
+            HelpMessage = "Storage account object",
+            ValueFromPipeline = true,
+            ParameterSetName = AccountObjectSingleParameterSet)]
         [Parameter(Mandatory = true,
             HelpMessage = "Storage account object",
             ValueFromPipeline = true,
@@ -75,15 +99,28 @@ namespace Microsoft.Azure.Commands.Management.Storage
 
         [Alias("N", "ShareName")]
         [Parameter(HelpMessage = "Share Name",
-            Mandatory = false)]
+            Mandatory = true,
+            ParameterSetName = AccountObjectSingleParameterSet)]
+        [Parameter(HelpMessage = "Share Name",
+            Mandatory = false,
+            ParameterSetName = AccountNameSingleParameterSet)]
         public string Name { get; set; }
 
+        [Parameter(Mandatory = false,
+            HelpMessage = "Include deleted shares, by default list shares won't include deleted shares",
+            ParameterSetName = AccountNameParameterSet)]
+        [Parameter(Mandatory = false,
+            HelpMessage = "Include deleted shares, by default list shares won't include deleted shares",
+            ParameterSetName = AccountObjectParameterSet)]
+        public SwitchParameter IncludeDeleted { get; set; }
+        
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
 
             switch (ParameterSetName)
             {
+                case AccountObjectSingleParameterSet:
                 case AccountObjectParameterSet:
                     this.ResourceGroupName = StorageAccount.ResourceGroupName;
                     this.StorageAccountName = StorageAccount.StorageAccountName;
@@ -95,7 +132,7 @@ namespace Microsoft.Azure.Commands.Management.Storage
                     this.Name = shareResource.ResourceName;
                     break;
                 default:
-                    // For AccountNameParameterSet, the ResourceGroupName and StorageAccountName can get from input directly
+                    // For AccountNameSingleParameterSet, the ResourceGroupName and StorageAccountName can get from input directly
                     break;
             }
 
@@ -109,10 +146,21 @@ namespace Microsoft.Azure.Commands.Management.Storage
             }
             else
             {
-                var Shares = this.StorageClient.FileShares.List(
+                ListSharesExpand? listSharesExpand = null;
+                if (this.IncludeDeleted.IsPresent)
+                {
+                    listSharesExpand = ListSharesExpand.Deleted;
+                }
+                IPage<FileShareItem> shares = this.StorageClient.FileShares.List(
                            this.ResourceGroupName,
-                           this.StorageAccountName);
-                WriteShareList(Shares);
+                           this.StorageAccountName,
+                           expand: listSharesExpand);
+                WriteShareList(shares);
+                while (shares.NextPageLink != null)
+                {
+                    shares = this.StorageClient.FileShares.ListNext(shares.NextPageLink);
+                    WriteShareList(shares);
+                }
             }
         }
     }